Move RowContextMenuCallbacks to viewModel
